+++ Source/Bugsnag.m
@@ -182,19 +182,23 @@ static NSMutableArray <id<BugsnagPlugin>> *registeredPlugins;
     }
 }
 
-+ (void)addAttribute:(NSString *)attributeName
-           withValue:(id)value
-       toTabWithName:(NSString *)tabName {
+/**
+ * Add custom data to send to Bugsnag with every exception. If value is nil,
+ * delete the current value for attributeName
+ */
++ (void)addMetadataToSection:(NSString *_Nonnull)section
+                         key:(NSString *_Nonnull)key
+                       value:(id _Nullable)value {
     if ([self bugsnagStarted]) {
-        [self.notifier.configuration.metadata addAttribute:attributeName
+        [self.notifier.configuration.metadata addAttribute:key
                                                  withValue:value
-                                             toTabWithName:tabName];
+                                             toTabWithName:section];
     }
 }
 
-+ (void)clearTabWithName:(NSString *)tabName {
++ (void)clearMetadataInSection:(NSString *)section {
     if ([self bugsnagStarted]) {
-        [self.notifier.configuration.metadata clearTab:tabName];
+        [self.notifier.configuration.metadata clearMetadataInSection:section];
     }
 }
 
@@ -298,6 +302,16 @@ static NSMutableArray <id<BugsnagPlugin>> *registeredPlugins;
         [[BSG_KSCrash sharedInstance]
                 setWriteBinaryImagesForUserReported:writeBinaryImagesForUserReported];
     }
+}
+
++ (NSMutableDictionary *)getMetadata:(NSString *)section {
+    return [[[self configuration] metadata] getMetadata:section];
+}
+
++ (id _Nullable )getMetadata:(NSString *_Nonnull)section
+                         key:(NSString *_Nonnull)key
+{
+    return [[[self configuration] metadata] getMetadata:section key:key];
 }
 
 @end