+++ examples/objective-c-ios/Bugsnag Test App/ViewController.m
@@ -67,7 +67,7 @@
         [NSJSONSerialization dataWithJSONObject:actuallyReallyJSON options:0 error:nil];
     } @catch (NSException *exception) {
         [Bugsnag notify:exception block:^(BugsnagEvent * _Nonnull report) {
-            report.metaData = @{@"tab": @{@"user": @"Bob Loblaw"}};
+            report.metadata = @{@"tab": @{@"user": @"Bob Loblaw"}};
         }];
     }