+++ CHANGELOG.md
@@ -25,6 +25,20 @@ Bugsnag Notifiers on other platforms.
 
 * Add a breadcrumb when Bugsnag first starts with the message "Bugsnag loaded"
   [#445](https://github.com/bugsnag/bugsnag-cocoa/pull/445)
+  
+* `Bugsnag.addAttribute:value:tab:` is now `Bugsnag.addMetadataToSection::key:value:`
+  [#454](https://github.com/bugsnag/bugsnag-cocoa/pull/454)
+  
+  *  `[Bugsnag clearTab:]` is now `[Bugsnag clearMetadataInSection:]` 
+     (Swift: `Bugsnag.clearMetadata(_ section)`)
+     [#457](https://github.com/bugsnag/bugsnag-cocoa/pull/457)
+
+* Added `Bugsnag.getMetadata(_ section)`.  The behaviour is: calling with a valid section
+  name will return the metadata for that section if it exists, or `nil` if it does not exist.  Other,
+  similar functionality (e.g. `BugsnagConfiguration.getTab()` has been renamed and
+  had usage aligned with this change.
+  [#459](https://github.com/bugsnag/bugsnag-cocoa/pull/459)
+  
 
 * Add a per-Event `apiKey` property.  This defaults to the global 
   `BugsnagConfiguration` value but can be overridden in event passed to the